refactor(pi): modularize user extension features
- group mode, status, thinking, and file tools under `user`
- extract shared file, mode, and completion helpers
- simplify secret policy messages and tool registration
