+++ files/pi/agent/extensions/user/lib/completions.ts
@@ -1,0 +1,9 @@
+import type { AutocompleteItem } from "@earendil-works/pi-tui";
+
+export function prefixCompletions(
+	items: readonly AutocompleteItem[],
+	prefix: string,
+): AutocompleteItem[] | null {
+	const filtered = items.filter((item) => item.value.startsWith(prefix));
+	return filtered.length > 0 ? filtered : null;
+}
