+++ files/pi/agent/extensions/user/lib/feature.ts
@@ -1,0 +1,13 @@
+import type { ExtensionAPI } from "@earendil-works/pi-coding-agent";
+
+/**
+ * A self-contained unit of the `user` extension (mode switching, status bar,
+ * thinking effort, file tools, ...). Each feature registers its own flags,
+ * commands, shortcuts, tools and event handlers against the pi extension API.
+ */
+export interface Feature {
+	/** Identifier for logging / ordering; not surfaced in pi's UI. */
+	readonly name: string;
+	/** Wire this feature into pi. Called once at extension load. */
+	register(pi: ExtensionAPI): void;
+}
