+++ files/pi/agent/extensions/user/lib/file/guard.ts
