+++ files/pi/agent/extensions/user/lib/secrets.ts
